[maintenance] Support priorities for tables in MM compaction
This commit add a feature to specify different priorities for table compaction.

In a Kudu cluster with thousands of tables, it's hard for a specified tablet's
maintenance OPs to be launched when their scores are not the highest, even if
the table the tablet belongs to is high priority for Kudu users. This patch
allow administators to specify different priorities for tables by gflags, these
maintenance OPs of these high priority tables have greater chance to be launched.

Change-Id: I3ea3b73505157678a8fb551656123b64e6bfb304