KUDU-3304: [Alter] Support to alter table's replication factor
In some cases we want to increase a table's replciation factor. For
example, convert a test table to production table, fix a table's RF
which is created by a buggy program.

This patch add a function to alter table's replication factor, supported
in CLI tool, and also add some unit tests.

Change-Id: I3aa2d5b12c508ba761fa9410ad1a465cf31fb7d7