[tool] Support to rewrite and print cmeta by batch
Both print and rewrite cmeta is operate on a single tablet, it will
cost a very long time if there are thousands of block containers
to rewrite. In some cases we need to rewrite a batch of raft configs
(e.g. 2 tservers permanently failed in a 3 tservers cluster).
This patch adds support to open FsManager once and print or rewrite
cmeta by batch.

Change-Id: I33dcbf5704bce4265eae0bed0f4296d17e352780