[tools] use thread pool for 'kudu local_replica copy_from_local'
In real world, we often use copy_from_local to copy a large
amount of tablets, we can pass a tablet id list to the tool
for convenient, and we can also gain profit of opening
filesystem only once to reduce time.

This patch also updates the related unit test.

Change-Id: I4de49e948cc1a686db5e1bf424470ca9e800ee36