[tools] Add 'kudu local_replica tmeta delete_rowsets' to delete rowsets from tablet
There are some use cases we need to delete rowsets from a tablet.
For example：
1. Some blocks are corrupted in a single node cluster, the server cannot be
   started. Note: some data will be lost in this case.
2. Some rowsets are fully deleted but the blocks can not be GCed (KUDU-3367).
   Note: no data will be lost in this case.

There is 'kudu pbc edit' CLI tool to achieve that, but it's error prone and
hard to operate when working with large amount of data.

This patch introduces a new CLI tool 'kudu local_replica tmeta delete_rowsets'
which makes removing rowsets from a tablet much easier.

Change-Id: If2cf9035babf4c3af4c238cebe8dcecd2c65848f