[Python] Add support for immutable column
This patch is a followup on: f20dcf57ad76e9b1bb57fe60b27ea3a8f02df233,
it introduces the immutable column type in the Python client.

There is a slight difference compared to the C++ API. In the C++ client
one can set two type of KuduColumnSpec: immutable or mutable. In the
Python client there is a single mutable ColumnSpec, which accepts a
boolean value. This is not a new type of difference between the clients,
for example the nullability works the same way, in terms of client
implementations.

Change-Id: I09bc3fe54ea7b40ccefa046c668e3622794b22e9
Reviewed-on: http://gerrit.cloudera.org:8080/19586
Tested-by: Kudu Jenkins
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>