[client] fix API non-compliance in KuduUpsertIgnore
Kudu C++ client API is still assumed to be C++98 compatible.  That's
an invariant that has been in place in the very beginning of 1.x version
history, so it should be kept as-is unless changing the major version.

Prior to this patch, client_examples-test.sh would fail while checking
a standalone compilation of kudu/client/write_op.h

This is a follow-up to ec3a9f75b6924a70ecbf08e3805228ad9b92b9f0.

Change-Id: I83f4e65bc3811ac701f5908beb480b39ad413641
Reviewed-on: http://gerrit.cloudera.org:8080/19765
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>