[fs] Simplify the data directory inheritance hierarchy
This patch doesn't introduce functional changes, but
only simplify the inheritance hierarchy of class Dir
to reduce the complex when extend the Dir class, for
example, when adding a sub-class of Dir to contain a
RocksDB instance, see http://gerrit.cloudera.org:8080/18569.

Because there are not sufficient reasons, so the
DataDir sub-class is removed and the implementation
functions are moved to the super class Dir.

Change-Id: I2d7059e32c7fd8444256ea031bebd153b318a5a3